Add lint.gradle and version.gradle and update some test gradle dependencies
The version.gradle has two funcion to generate the versionCode and
versinName based on the respository commit count.

Lint.gradle setup the KtLint on this project to make sure everything
is ok with the default Kotlin code style.

Added the nhaarman library that will be useful to write tests later.